+++ lib/teamspeak3.rb
@@ -1,7 +1,7 @@
 require 'net/telnet'
 
 module TeamSpeak3
-  VERSION = "0.0.5"
+  VERSION = "0.0.6"
 end
 
 # exceptions shall be required first